+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Fix connecteur dsp2 civicrm. Refactoring du connecteur, pour plus de cohérence avec les autres connecteurs.
 * On stocke le statut distant des transactions.
 * Fix balises html fermantes manquantes.
+* Conversions adhésions/dons.
+* Fix: pas d'erreur quand un connecteur n'est pas trouvé.
 
 ## 0.0.4